creation of base class, pull out checkLocation
